refactorisation of the code by moving every functions in different files + upgrading of the print in the main
