+++ data_WaTor.csv
@@ -1,314 +1,73 @@
 
-chronon,nb_fish,nb_shark
-0,200,25
-1,188,25
-2,358,16
-3,351,16
-4,686,14
-5,682,13
-6,1197,12
-7,1241,24
-8,1535,24
-9,1551,24
-10,1575,24
-11,1576,24
-12,1576,24
-13,1576,24
-14,1552,48
-15,1548,48
-16,1548,48
-17,1547,48
-18,1549,48
-19,1549,48
-20,1551,48
-21,1504,96
-22,1480,96
-23,1447,96
-24,1448,96
-25,1438,96
-26,1452,96
-27,1444,96
-28,1380,190
-29,1320,190
-30,1249,187
-31,1195,185
-32,1136,178
-33,1100,175
-34,1064,168
-35,992,324
-36,869,314
-37,777,254
-38,733,236
-39,687,189
-40,643,159
-41,624,147
-42,584,274
-43,512,256
-44,458,192
-45,436,177
-46,427,134
-47,447,105
-48,461,89
-49,516,168
-50,505,165
-51,537,132
-52,603,120
-53,651,104
-54,750,89
-55,804,83
-56,889,160
-57,897,155
-58,922,134
-59,975,129
-60,1018,121
-61,1096,107
-62,1136,104
-63,1131,204
-64,1091,203
-65,1049,194
-66,1028,188
-67,1031,176
-68,1033,166
-69,1057,159
-70,1001,290
-71,935,278
-72,843,241
-73,812,231
-74,757,215
-75,750,199
-76,711,189
-77,680,350
-78,559,341
-79,527,265
-80,475,233
-81,514,183
-82,492,151
-83,558,128
-84,552,224
-85,570,206
-86,553,143
-87,611,135
-88,640,117
-89,737,104
-90,783,94
-91,884,176
-92,879,171
-93,958,152
-94,984,148
-95,1046,143
-96,1100,135
-97,1143,132
-98,1135,258
-99,1085,252
-100,1059,228
-101,1019,215
-102,1009,194
-103,989,175
-104,993,164
-105,914,312
-106,783,303
-107,686,268
-108,638,259
-109,617,211
-110,616,166
-111,623,135
-112,625,232
-113,577,219
-114,583,163
-115,596,151
-116,639,116
-117,658,97
-118,707,91
-119,714,166
-120,723,163
-121,714,127
-122,752,117
-123,773,93
-124,832,85
-125,868,84
-126,904,162
-127,895,162
-128,936,148
-129,938,142
-130,1009,128
-131,1038,112
-132,1118,103
-133,1115,200
-134,1093,196
-135,1036,169
-136,1029,162
-137,1014,152
-138,1033,136
-139,1036,129
-140,1020,253
-141,974,244
-142,922,211
-143,914,205
-144,899,189
-145,882,178
-146,909,172
-147,866,333
-148,793,325
-149,722,269
-150,685,247
-151,662,212
-152,650,187
-153,658,161
-154,635,298
-155,578,285
-156,536,219
-157,523,196
-158,505,172
-159,523,147
-160,524,134
-161,541,242
-162,493,230
-163,527,175
-164,547,163
-165,619,127
-166,682,103
-167,777,87
-168,834,158
-169,876,147
-170,898,130
-171,947,127
-172,996,120
-173,1061,116
-174,1109,110
-175,1119,204
-176,1091,201
-177,1067,186
-178,1070,176
-179,1071,164
-180,1064,148
-181,1074,141
-182,1000,282
-183,922,279
-184,826,258
-185,803,246
-186,757,204
-187,779,180
-188,764,158
-189,761,276
-190,661,269
-191,615,230
-192,557,215
-193,558,194
-194,533,165
-195,576,142
-196,562,256
-197,552,242
-198,511,189
-199,543,170
-200,543,138
-201,589,119
-202,633,110
-203,665,198
-204,686,187
-205,701,155
-206,747,148
-207,795,141
-208,865,131
-209,919,124
-210,933,233
-211,896,229
-212,893,198
-213,894,183
-214,921,155
-215,930,135
-216,966,125
-217,932,238
-218,873,234
-219,820,211
-220,832,201
-221,822,161
-222,866,145
-223,867,139
-224,884,254
-225,809,251
-226,800,216
-227,766,202
-228,779,177
-229,779,164
-230,816,151
-231,793,284
-232,778,281
-233,750,228
-234,740,213
-235,735,188
-236,758,165
-237,782,143
-238,786,248
-239,775,236
-240,763,177
-241,784,160
-242,795,138
-243,825,128
-244,851,125
-245,839,238
-246,811,235
-247,789,190
-248,789,175
-249,809,144
-250,824,137
-251,865,124
-252,845,228
-253,813,223
-254,768,194
-255,753,189
-256,732,175
-257,749,160
-258,766,146
-259,770,254
-260,741,244
-261,738,194
-262,753,173
-263,767,147
-264,797,131
-265,839,119
-266,852,218
-267,816,212
-268,811,176
-269,817,164
-270,831,143
-271,863,135
-272,914,118
-273,913,216
-274,896,209
-275,846,188
-276,870,180
-277,854,154
-278,920,143
-279,929,135
-280,973,256
-281,900,255
-282,900,244
-283,857,240
-284,888,225
-285,867,214
-286,903,208
-287,822,388
-288,715,384
-289,579,324
-290,517,298
-291,454,253
-292,435,198
-293,427,169
-294,420,291
-295,336,279
-296,312,207
-297,306,187
-298,345,135
-299,372,100
-300,437,77
-301,480,128
-302,538,125
-303,576,99
-304,664,93
-305,744,85
-306,848,77
-307,954,77
-308,1043,154
-309,1086,153
-310,1138,148
-311,1171,148
+chronon,nb_fish,nb_shark,average_age_fish,average_age_shark,nb_descendants_fish,nb_descendants_shark,average_distance_fish,average_distance_shark,max_nb_descendants_fish,max_nb_descendants_shark,max_distance_fish,max_distance_shark
+0,200,25,0,0,0,0,0,0,0,0,0,0
+1,193,25,0,1,0,0,1,1,0,0,1,1
+2,378,10,0,2,0,0,1,2,1,0,3,2
+3,374,10,0,3,0,0,2,3,1,0,4,3
+4,729,7,0,4,0,0,2,4,2,0,6,4
+5,734,7,0,5,0,0,3,5,2,0,7,5
+6,1253,6,0,6,0,0,2,6,3,0,9,6
+7,1312,12,0,3,0,0,3,4,3,1,10,8
+8,1549,12,0,4,0,0,2,5,4,1,12,9
+9,1566,12,0,5,0,0,2,6,4,1,13,10
+10,1588,12,0,6,0,0,2,7,5,1,15,11
+11,1587,12,0,7,0,0,2,8,5,1,15,12
+12,1588,12,0,8,0,0,2,9,6,1,16,13
+13,1588,12,0,9,0,0,2,10,6,1,16,14
+14,1576,24,0,5,0,0,2,6,6,2,16,16
+15,1575,24,0,6,0,0,2,7,7,2,18,17
+16,1575,24,0,7,0,0,2,8,7,2,19,18
+17,1575,24,0,8,0,0,2,9,8,2,21,19
+18,1576,24,0,9,0,0,2,10,5,2,15,20
+19,1576,24,0,10,0,0,2,11,5,2,15,21
+20,1575,24,0,11,0,0,2,12,5,2,15,22
+21,1552,48,0,6,0,0,2,7,5,3,15,24
+22,1544,48,0,7,0,0,2,8,5,3,15,25
+23,1534,48,0,8,0,0,2,9,5,3,15,26
+24,1538,47,0,9,0,0,2,9,5,3,15,27
+25,1545,47,0,10,0,0,2,10,5,3,15,28
+26,1548,47,0,11,0,0,2,11,6,3,17,29
+27,1547,47,0,12,0,0,2,12,5,3,15,30
+28,1506,94,0,6,0,0,2,7,5,4,15,32
+29,1486,94,0,7,0,0,2,8,5,4,15,33
+30,1461,93,0,8,0,0,2,9,6,4,17,34
+31,1456,92,0,9,0,0,2,10,6,4,17,35
+32,1456,92,0,10,0,0,2,11,6,4,17,36
+33,1462,92,0,11,0,0,2,12,7,4,19,37
+34,1469,92,0,12,0,0,2,13,6,4,15,38
+35,1401,184,0,6,0,0,2,7,6,5,16,40
+36,1337,184,0,7,0,0,2,8,7,5,18,41
+37,1266,183,0,8,0,0,2,9,7,5,19,42
+38,1230,182,0,9,0,0,2,10,5,5,15,43
+39,1192,174,0,11,0,1,2,12,6,5,16,44
+40,1173,167,0,12,0,1,2,13,6,5,17,45
+41,1146,165,0,13,0,1,2,14,6,5,18,46
+42,1085,320,0,7,0,1,2,8,7,6,20,48
+43,943,315,0,8,0,1,2,9,7,6,20,49
+44,835,275,0,9,0,1,2,10,8,6,22,50
+45,762,262,0,10,0,1,2,11,8,6,23,51
+46,713,225,0,11,0,1,2,12,5,6,16,52
+47,658,189,0,12,0,1,2,13,6,6,18,53
+48,646,171,0,13,0,1,2,14,6,6,19,54
+49,589,305,0,7,0,1,2,8,6,7,18,56
+50,524,287,0,7,0,0,2,8,6,7,18,57
+51,448,206,0,9,0,1,3,10,6,6,19,50
+52,444,186,0,10,0,1,2,10,7,6,21,51
+53,422,156,0,11,0,1,3,12,7,6,22,52
+54,458,131,0,12,0,1,3,13,8,6,24,53
+55,477,115,0,13,0,1,3,14,8,6,25,54
+56,549,206,0,7,0,1,3,8,9,7,27,56
+57,545,197,0,8,1,1,3,9,9,7,28,57
+58,614,154,0,10,1,1,3,11,10,7,30,58
+59,642,140,0,10,1,1,3,11,10,7,31,59
+60,769,117,0,11,1,1,3,12,11,6,33,52
+61,830,99,0,12,1,1,3,13,11,6,34,53
+62,948,83,0,12,1,0,3,13,12,5,36,46
+63,960,150,0,6,1,0,3,7,12,6,36,48
+64,1014,142,0,7,0,0,3,8,10,6,30,49
+65,1018,121,0,9,0,1,3,10,10,6,30,50
+66,1078,119,0,10,0,1,3,11,10,6,30,51
+67,1095,114,0,11,0,1,2,12,10,6,30,52
+68,1154,113,0,12,0,1,2,13,10,6,30,53
+69,1179,110,0,13,0,1,2,14,10,6,30,54
+70,1177,220,0,7,0,1,2,8,10,7,30,56
